+++ src/main/java/lass9436/question/servlet/QuestionServlet.java
@@ -1,5 +1,6 @@
 package lass9436.question.servlet;
 
+import java.io.BufferedReader;
 import java.io.IOException;
 
 import jakarta.servlet.ServletConfig;
@@ -12,6 +13,8 @@ import lass9436.question.model.Question;
 import lass9436.question.model.QuestionRepository;
 import lass9436.user.model.User;
 import lass9436.user.model.UserRepository;
+import lass9436.utils.JsonUtil;
+import org.json.JSONObject;
 
 @WebServlet("/questions")
 public class QuestionServlet extends HttpServlet {
@@ -39,6 +42,26 @@ public class QuestionServlet extends HttpServlet {
 		if (user != null && user.getName().equals(writer)){
 			questionRepository.save(new Question(user.getUserSeq(), writer, title, contents));
 			resp.sendRedirect("/");
+			return;
+		}
+
+		resp.sendError(HttpServletResponse.SC_FORBIDDEN, "User not authorized.");
+	}
+
+	@Override
+	protected void doPut(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
+		JSONObject json = JsonUtil.parseJson(req);
+		long seq = json.getLong("seq");
+		String writer = json.getString("writer");
+		String title = json.getString("title");
+		String contents = json.getString("contents");
+
+		String userId = (String)req.getSession().getAttribute("userId");
+		User user = userRepository.findByUserId(userId);
+
+		if (user != null && user.getName().equals(writer)) {
+			questionRepository.save(new Question(seq, user.getUserSeq(), writer, title, contents));
+			resp.setStatus(HttpServletResponse.SC_OK);
 			return;
 		}
